pyproject.toml
[tool.poetry]
name = "answer-bot"
packages = [{include = "agent"}]
version = "0.1.0"
description = ""
authors = ["EgorBodrov <bodroff.e.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.7"
langchain-core = "^0.3.17"
langchain-community = "^0.3.7"
langgraph = "^0.2.45"
python-dotenv = "^1.0.1"
gigachat = "^0.1.35"
langchain-openai = "^0.2.8"


[tool.poetry.group.database.dependencies]
torch = {version = "^2.5.1+cu121", source = "pytorch-gpu"}
qdrant-client = "^1.12.1"
sentence-transformers = "^3.3.0"


[tool.poetry.group.backend.dependencies]
fastapi = "^0.115.5"
uvicorn = "^0.32.0"


[tool.poetry.group.frontend.dependencies]
streamlit = "^1.40.1"


[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
